AI/ImageHeader/requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=9.5.0
torchvision>=0.15.0
